Add encoder config file

+++ tmk_core/protocol/nrf/keycode_str_converter.h
@@ -15,8 +15,8 @@ typedef struct {
   const char* key_strings;
 } key_string_map_t;
 
-uint8_t quantum_keycode2str(uint16_t qk, char* str, uint8_t len);
-uint8_t quantum_keycode2str_locale(uint16_t qk, char* str, uint8_t len,
+uint8_t quantum_keycode2str(uint16_t qk, char* str, uint32_t len);
+uint8_t quantum_keycode2str_locale(uint16_t qk, char* str, uint32_t len,
   KEYMAP_LOCALE locale, bool use_ascii);
 uint16_t get_kc_strlen(const char* str, uint16_t len);
 uint16_t get_inner_element(const char** pp_str, uint16_t len);
